+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Add new functions sample_gp() and sample_magmaclust() to sample from the posteriors predictions of GPs, Magma, and MagmaClust
 * Propose a new visualisation based on posterior samples instead of Credible Intervals for both plot_gp() and plot_magmaclust()
+* Allow pred_*() functions to be used without providing the 'data' argument by returning the mean processes (or priors) as generic predictions. 
 
 ## Minor 
 
@@ -11,6 +12,7 @@
 * Fix a bug when the 'Reference' column is present when using plot_gp() 
 * Fix the '\docType{package}' bug in roxygen2
 * Add 'plot_magma()' as a duplicated name for 'plot_gp()'
+* Remove the useless 'Reference' column in the prediction of the mean processes
 
 # MagmaClustR 1.1.2